[ AMSD-63 ] Fixed the flight crew entity CRUD needed operations

+++ src/main/java/com/internationalairport/airportmanagementsystem/dtos/post/PostParkingDto.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 public record PostParkingDto(
     String location,
     Integer capacity,
-    BigDecimal rate
+    Double rate
 ){
 
 }
